refactor(user service): removed unused methods. Moved getpermissions so separate service
This is to make the user service more testable and fakeable
