+++ src/Sepes.Infrastructure/Service/Interface/IUserPermissionService.cs
@@ -1,0 +1,10 @@
+﻿using Sepes.Infrastructure.Dto;
+using System.Threading.Tasks;
+
+namespace Sepes.Infrastructure.Service.Interface
+{
+    public interface IUserPermissionService
+    {
+        Task<UserPermissionDto> GetUserPermissionsAsync();
+    }
+}
