--- src/Sepes.Tests/Mocks/PrincipalServiceMock.cs
@@ -1,16 +1,0 @@
-﻿using Sepes.Infrastructure.Interface;
-using System;
-using System.Collections.Generic;
-using System.Security.Principal;
-using System.Text;
-
-namespace Sepes.Tests.Mocks
-{
-    public class PrincipalServiceMock : IPrincipalService
-    {
-        public IPrincipal GetPrincipal()
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
